pyproject.toml
[project]
name = "ledger-nano-s-iota-installer"
version = "0.9.2"
description = "Helper tool for installing IOTA on Ledger Nano S"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ecdsa>=0.19.1",
    "ledgerwallet>=0.1.4",
    "protobuf>=6.30.2",
    "pygithub>=2.6.1",
    "requests>=2.32.3",
    "setuptools>=80.1.0",
]
